src/email/requirements-email.txt
python-dotenv
grpcio
grpcio-tools
protobuf